move static files
